firebase edistystä
