+++ src/components/SignUp.js
@@ -2,7 +2,7 @@ import React from 'react'
 import logo from '../img/WTlogo_stacked_white_bordered.png'
 import google from '../img/google_logo.png'
 import { getDatabase, ref, set } from "firebase/database";
-import { firebaseConfig } from "../index.js"
+import { firebaseConfig } from "../firebase.js"
 
 const db = getDatabase(firebaseConfig.appId);
