+++ src/firebase.js
@@ -1,0 +1,22 @@
+import firebase from 'firebase/compat/app'
+import 'firebase/compat/auth'
+import 'firebase/compat/database'
+import 'firebase/compat/analytics'
+import { getDatabase } from "firebase/database";
+
+
+export const firebaseConfig = {
+    apiKey: process.env.REACT_APP_FIREBASE_APP_API_KEY,
+    authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
+    projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
+    storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
+    messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
+    appId: process.env.REACT_APP_FIREBASE_APP_ID,
+    measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENT_ID,
+    databaseURL: process.env.REACT_APP_FIREBASE_DATABASE_URL
+  }
+
+  firebase.initializeApp(firebaseConfig)
+  const analytics = firebase.analytics()
+
+  export default firebase;
