+++ src/index.js
@@ -4,28 +4,15 @@ import { BrowserRouter } from 'react-router-dom'
 import './index.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
-import firebase from 'firebase/compat/app'
-import 'firebase/compat/auth'
-import 'firebase/compat/database'
-import 'firebase/compat/analytics'
 import 'react-toggle/style.css'
-import { getDatabase } from "firebase/database";
+import firebase from "./firebase.js"
 
-export const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_APP_API_KEY,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_APP_ID,
-  measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENT_ID,
-  databaseURL: process.env.REACT_APP_FIREBASE_DATABASE_URL
-}
+
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'))
-firebase.initializeApp(firebaseConfig)
-const analytics = firebase.analytics()
+
+
 root.render(
   <React.StrictMode>
     <BrowserRouter>
